modify: database 관련 수정
* Stamp에 primary key로 쓰일 프로퍼티 추가
* RoomDatabase 구현체에 TypeConverter 제공
* addStamps가 리턴하는 id의 타입을 Int -> Long으로 수정

+++ data/src/main/java/com/thk/data/database/SevenDaysDao.kt
@@ -25,7 +25,7 @@ interface StampDao {
     suspend fun getStamps(challengeId: Long): List<Stamp>
 
     @Insert(onConflict = REPLACE)
-    suspend fun addStamps(vararg stamp: Stamp): List<Int>
+    suspend fun addStamps(vararg stamp: Stamp): List<Long>
 
     @Update
     suspend fun updateStampChecked(stamp: Stamp)
